Move reporting out of the upgrade job.
Reporting the results did not work properly, because it was done inside
the upgrade job, which of course would no longer run when the
application was restarted. Therefore, the upgrade-spp.sh shell script
now uses rails runner to invoke a new helper method _after_ the
application has been restarted, and this helper will take care of
sending out an e-mail.